Add files via upload
Basic Concept of Arrayed Waveguide Grating, Integrated Circuit by Jack Kilby (1958), Photonic Integrated Circuit Structure, (a) Cross section of a typical silicon photonic wire waveguide. (b) Scheme of the direct silicon waveguide crossing on silicon-on-insulator (SOI) platform. (c) The ray model of the fundamental mode in the waveguide where 𝜃𝑝 is the light projection angle and 𝜃ℎ is the horn angle of the optical waveguide.